+++ cmd/server/main.go
@@ -17,7 +17,7 @@ import (
 )
 
 const (
-	Version = "3.2-beta1"
+	Version = "3.2-beta2"
 )
 
 var (